backend/requirements.txt
django==5.2.8
djangorestframework==3.16.1
python-dotenv==1.2.1
psycopg2-binary==2.9.11
djangorestframework-simplejwt==5.5.1
django-cors-headers
drf-spectacular
drf-spectacular-sidecar
django-redis
redis
celery==5.3.6
eventlet==0.33.3
cryptography==41.0.7
requests==2.31.0
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2

# testing dependencies
pytest==8.0.0
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==22.0.0
